Late night hacking

+++ db/migrate/20171022045054_create_angles.rb
@@ -2,7 +2,7 @@ class CreateAngles < ActiveRecord::Migration[5.1]
   def change
     create_table :angles do |t|
       t.string :name
-      t.decimal :pivot
+      t.integer :pivot
       t.decimal :zoom_x, default: 0.0
       t.decimal :zoom_y, default: 0.0
       t.decimal :zoom_w, default: 1.0
